Order: Moving js to components folders and removing default manifest of react create app
